Extracted region tests from entry tests
