+++ test-plugins/org.yakindu.sct.model.sgraph.test/.settings/org.eclipse.core.resources.prefs
@@ -1,3 +1,3 @@
 eclipse.preferences.version=1
-encoding//testmodels/entry/RegionCantBeEnteredUsingShallowHistory.sct=UTF-8
+encoding//testmodels/region/RegionCantBeEnteredUsingShallowHistory.sct=UTF-8
 encoding/<project>=UTF-8
