+++ test-plugins/org.yakindu.sct.model.sgraph.test/src/org/yakindu/sct/model/sgraph/test/validation/EntryValidatorTest.java
@@ -24,12 +24,15 @@ import org.yakindu.sct.model.sgraph.Statechart;
 import org.yakindu.sct.model.sgraph.Transition;
 import org.yakindu.sct.model.sgraph.test.util.SGraphJavaValidatorTester;
 import org.yakindu.sct.model.sgraph.validation.EntryValidator;
+import org.yakindu.sct.model.sgraph.validation.RegionValidator;
 
 import com.google.inject.Inject;
 
 /**
  * 
  * Tests for {@link EntryValidator}
+ * 
+ * TODO: Test composite state regions
  *
  */
 public class EntryValidatorTest extends AbstractSGraphValidatorTest {
@@ -128,24 +131,6 @@ public class EntryValidatorTest extends AbstractSGraphValidatorTest {
 	}
 
 
-	/**
-	 * Tests a scenario where no issues for an exit nodes exists.
-	 */
-	@Test
-	public void cleanExit() {
-		State state = createState();
-		Region subRegion = factory.createRegion();
-		state.getRegions().add(subRegion);
-		Exit exit = factory.createExit();
-		subRegion.getVertices().add(exit);
-		State s = factory.createState();
-		subRegion.getVertices().add(s);
-		Transition t = factory.createTransition();
-		t.setTarget(exit);
-		t.setSource(s);
-
-		tester.validate(exit).assertOK();
-	}
 
 	@Test
 	public void disallowTrigger() {
@@ -158,40 +143,5 @@ public class EntryValidatorTest extends AbstractSGraphValidatorTest {
 		tester.validate(entry).assertError(ENTRY_TRANSITIONS_NO_TRIGGER_ON_OUT_CODE);
 	}
 
-	@Test
-	public void checkOnlyOneDefaultEntryPermitted() {
-		State state = createState();
-		Region region = ((Region) state.eContainer());
-		Entry entry = factory.createEntry();
-		region.getVertices().add(entry);
-		Entry entry2 = factory.createEntry();
-		entry2.setName("default");
-		region.getVertices().add(entry2);
-		createTransition(entry, state);
-		createTransition(entry2, state);
-		tester.validate(entry).assertError(REGION_MULTIPLE_DEFAULT_ENTRIES_CODE);
-	}
-
-	@Test
-	public void regionCantBeEnteredUsingShallowHistory() {
-		Statechart statechart = loadStatechart("RegionCantBeEnteredUsingShallowHistory.sct");
-		AssertableDiagnostics result = tester.validate(statechart);
-		result.assertDiagnosticsCount(2);
-
-		result.assertAny(
-				AssertableDiagnostics.errorCode(REGION_CANT_BE_ENTERED_USING_SHALLOW_HISTORY_NO_DEFAULT_ENTRY_CODE));
-		result.assertAny(AssertableDiagnostics
-				.errorCode(REGION_CANT_BE_ENTERED_USING_SHALLOW_HISTORY_NON_CONNECTED_DEFAULT_ENTRY_CODE));
-	}
-
-	@Test
-	public void validTransitionToInnerChoice() {
-		Statechart statechart = loadStatechart("ValidTransitionToInnerChoice.sct");
-		tester.validate(statechart).assertOK();
-	}
-
-	@Override
-	protected Statechart loadStatechart(String path) {
-		return super.loadStatechart("entry/" + path);
-	}
+	
 }
