+++ test-plugins/org.yakindu.sct.model.sgraph.test/testmodels/region/RegionCantBeEnteredUsingShallowHistory.sct
