+++ test-plugins/org.yakindu.sct.model.sgraph.test/testmodels/region/ValidTransitionToInnerChoice.sct
